pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pheme"
version = "0.1.0"
description = "Universal communication layer for agent-to-human notifications"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "apprise>=1.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.setuptools.packages.find]
include = ["server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
